Prepare module structure to adding more queries
